requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncio-throttle==1.0.2
schedule==1.2.0
pydantic==2.5.2
redis==5.0.1
openai==1.3.7
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
motor==3.3.2
pymongo==4.6.0
# Парсинг и автоматизация
selenium==4.15.2
playwright==1.40.0
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
lxml==4.9.3
# Прокси и обход блокировок
requests[socks]==2.31.0
pysocks==1.7.1
# Кэширование и очереди
celery==5.3.4
redis==5.0.1
